requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.10
coverage==6.2
fastjsonschema==2.15.3
findspark==2.0.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
py4j==0.10.9.2
pyparsing==3.0.6
pyspark==3.2.0
pytest==6.2.5
pytest-cov==3.0.0
pytest-mock==3.6.1
pytest-spark==0.6.0
PyYAML==6.0
requests==2.27.1
toml==0.10.2
urllib3==1.26.8
